db changes + login

+++ wdpr_her/backend/api/Data/Bedrijf.cs
@@ -3,7 +3,7 @@
 public class Bedrijf : Gebruiker
 {
     public string Naam { get; set; }
-    public int Kvk { get; set; }
+    public long Kvk { get; set; }
     public string? Locatie { get; set; }
     public string? Website { get; set; } // In ERR Link
 }
